proofread literally everything: Improve doc, fix random things, add preconditions

+++ test-util/src/main/kotlin/org/futurerobotics/temporaryname/TestMathUtil.kt
@@ -6,7 +6,14 @@ import org.futurerobotics.temporaryname.math.angleNorm
 import org.futurerobotics.temporaryname.math.distTo
 import kotlin.math.abs
 
-const val ADJUST_THRESH: Int = 10
+/**
+ * The threshold for adjustment in [errorTo] functions
+ */
+const val ADJUST_THRESH: Double = 10.0
+
+/**
+ * Gets a reasonable value for the error from this vector to that vector.
+ */
 infix fun Vector2d.errorTo(that: Vector2d): Double {
     val thisLen = this.length
     val otherLen = that.length
@@ -18,6 +25,9 @@ infix fun Vector2d.errorTo(that: Vector2d): Double {
     }
 }
 
+/**
+ * Gets a reasonable value for the error from this value to that value.
+ */
 infix fun Double.errorTo(that: Double): Double {
     val total = abs(this) + abs(that)
     val err = this distTo that
@@ -27,5 +37,8 @@ infix fun Double.errorTo(that: Double): Double {
     }
 }
 
+/**
+ * Gets a reasonable value for the error from this pose to that pose.
+ */
 infix fun Pose2d.errorTo(that: Pose2d): Double =
     kotlin.math.max(vec errorTo that.vec, angleNorm(heading distTo that.heading) / 3)
